Update coding standards
Because I upgraded the version of "SlevomatCodingStandard," I had to
upgrade specific rules that would break backward compatibility. Since
that was already the case, I added a couple more rules to improve the
coding standards.

I also bumped the support to PHP 8.1 since the support PHP will drop the
support for version 8.0 soon.

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>